+++ expo-quest-notifications/android/build.gradle
@@ -7,9 +7,9 @@ plugins {
 group = 'host.exp.exponent'
 version = '0.32.11'
 
-def horizonEnabled = providers.gradleProperty("horizonEnabled")
-    .map { it.toBoolean() }
-    .getOrElse(false)
+expoModule {
+  canBePublished false
+}
 
 android {
   namespace "expo.modules.notifications"
@@ -34,16 +34,10 @@ android {
     }
   }
 
-  sourceSets {
-    main {
-      java {
-       if (horizonEnabled) {
-         srcDirs = ['src/main/java', 'src/horizon/java']
-       } else {
-         srcDirs = ['src/main/java', 'src/android/java']
-       }
-      }
-    }
+  flavorDimensions += "device"
+  productFlavors {
+    quest { dimension "device" }
+    mobile { dimension "device" }
   }
 }
 
@@ -54,10 +48,8 @@ dependencies {
   implementation 'androidx.lifecycle:lifecycle-common-java8:2.2.0'
   implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3'
 
-  if (horizonEnabled) {
-    implementation("com.meta.horizon.platform.ovr:android-platform-sdk:77")
-    implementation files('libs/android-platform-sdk-v79.aar')
-  }
+  questApi("com.meta.horizon.platform.ovr:android-platform-sdk:77")
+  questApi files('libs/android-platform-sdk-v79.aar')
 
   // release notes in https://firebase.google.com/support/release-notes/android - cmd + f "Cloud Messaging version"
   implementation 'com.google.firebase:firebase-messaging:24.0.1'
@@ -70,8 +62,6 @@ dependencies {
   }
 }
 
-if (horizonEnabled) {
-  configurations.all {
-    exclude group: "com.meta.horizon.platform.ovr", module: "android-platform-sdk"
-  }
+configurations.all {
+  exclude group: "com.meta.horizon.platform.ovr", module: "android-platform-sdk"
 }
