+++ expo-quest/plugin/src/constants.ts
@@ -1,0 +1,127 @@
+/**
+ * List of Android permissions that are prohibited on Meta Quest devices.
+ * These permissions will be automatically blocked/removed from the manifest when building for Quest.
+ * 
+ * Source: Meta Quest Platform Requirements
+ * https://developer.oculus.com/documentation/native/android/mobile-publishing-guide/
+ */
+export const PROHIBITED_PERMISSIONS = [
+  "ACCEPT_HANDOVER",
+  "ACCESS_BACKGROUND_LOCATION",
+  "ACCESS_CHECKIN_PROPERTIES",
+  "ACCESS_LOCATION_EXTRA_COMMANDS",
+  "ACCESS_NOTIFICATION_POLICY",
+  "ACCOUNT_MANAGER",
+  "ACTIVITY_RECOGNITION",
+  "ADD_VOICEMAIL",
+  "ANSWER_PHONE_CALLS",
+  "BIND_ACCESSIBILITY_SERVICE",
+  "BIND_APPWIDGET",
+  "BIND_AUTOFILL_SERVICE",
+  "BIND_CALL_REDIRECTION_SERVICE",
+  "BIND_CARRIER_MESSAGING_CLIENT_SERVICE",
+  "BIND_CARRIER_MESSAGING_SERVICE",
+  "BIND_CARRIER_SERVICES",
+  "BIND_CHOOSER_TARGET_SERVICE",
+  "BIND_CONDITION_PROVIDER_SERVICE",
+  "BIND_CONTROLS",
+  "BIND_DEVICE_ADMIN",
+  "BIND_DREAM_SERVICE",
+  "BIND_INCALL_SERVICE",
+  "BIND_INPUT_METHOD",
+  "BIND_MIDI_DEVICE_SERVICE",
+  "BIND_NFC_SERVICE",
+  "BIND_NOTIFICATION_LISTENER_SERVICE",
+  "BIND_PRINT_SERVICE",
+  "BIND_QUICK_ACCESS_WALLET_SERVICE",
+  "BIND_QUICK_SETTINGS_TILE",
+  "BIND_REMOTEVIEWS",
+  "BIND_SCREENING_SERVICE",
+  "BIND_TELECOM_CONNECTION_SERVICE",
+  "BIND_TEXT_SERVICE",
+  "BIND_TV_INPUT",
+  "BIND_VISUAL_VOICEMAIL_SERVICE",
+  "BIND_VOICE_INTERACTION",
+  "BIND_VR_LISTENER_SERVICE",
+  "BIND_WALLPAPER",
+  "BLUETOOTH_PRIVILEGED",
+  "BODY_SENSORS",
+  "BROADCAST_PACKAGE_REMOVED",
+  "BROADCAST_SMS",
+  "BROADCAST_WAP_PUSH",
+  "CALL_PHONE",
+  "CALL_PRIVILEGED",
+  "CAPTURE_AUDIO_OUTPUT",
+  "CHANGE_COMPONENT_ENABLED_STATE",
+  "CHANGE_CONFIGURATION",
+  "CLEAR_APP_CACHE",
+  "CONTROL_LOCATION_UPDATES",
+  "DELETE_CACHE_FILES",
+  "DELETE_PACKAGES",
+  "DIAGNOSTIC",
+  "DUMP",
+  "FACTORY_TEST",
+  "GET_ACCOUNTS",
+  "GET_ACCOUNTS_PRIVILEGED",
+  "INSTALL_LOCATION_PROVIDER",
+  "INSTALL_PACKAGES",
+  "INSTANT_APP_FOREGROUND_SERVICE",
+  "LOADER_USAGE_STATS",
+  "LOCATION_HARDWARE",
+  "MANAGE_DOCUMENTS",
+  "MANAGE_MEDIA",
+  "MANAGE_ONGOING_CALLS",
+  "MASTER_CLEAR",
+  "MEDIA_CONTENT_CONTROL",
+  "MODIFY_PHONE_STATE",
+  "MOUNT_FORMAT_FILESYSTEMS",
+  "MOUNT_UNMOUNT_FILESYSTEMS",
+  "PACKAGE_USAGE_STATS",
+  "PROCESS_OUTGOING_CALLS",
+  "QUERY_ALL_PACKAGES",
+  "READ_CALENDAR",
+  "READ_CALL_LOG",
+  "READ_CONTACTS",
+  "READ_INPUT_STATE",
+  "READ_LOGS",
+  "READ_PHONE_NUMBERS",
+  "READ_PHONE_STATE",
+  "READ_PRECISE_PHONE_STATE",
+  "READ_SMS",
+  "READ_VOICEMAIL",
+  "REBOOT",
+  "RECEIVE_MMS",
+  "RECEIVE_SMS",
+  "RECEIVE_WAP_PUSH",
+  "REQUEST_DELETE_PACKAGES",
+  "REQUEST_INSTALL_PACKAGES",
+  "SEND_RESPOND_VIA_MESSAGE",
+  "SEND_SMS",
+  "SET_ALWAYS_FINISH",
+  "SET_ANIMATION_SCALE",
+  "SET_DEBUG_APP",
+  "SET_PROCESS_LIMIT",
+  "SET_TIME",
+  "SET_TIME_ZONE",
+  "SIGNAL_PERSISTENT_PROCESSES",
+  "SMS_FINANCIAL_TRANSACTIONS",
+  "START_FOREGROUND_SERVICES_FROM_BACKGROUND",
+  "START_VIEW_PERMISSION_USAGE",
+  "STATUS_BAR",
+  "SYSTEM_ALERT_WINDOW",
+  "UNINSTALL_SHORTCUT",
+  "UPDATE_DEVICE_STATS",
+  "USB_CAMERA",
+  "USE_ICC_AUTH_WITH_DEVICE_IDENTIFIER",
+  "USE_SIP",
+  "UWB_RANGING",
+  "WRITE_APN_SETTINGS",
+  "WRITE_CALENDAR",
+  "WRITE_CALL_LOG",
+  "WRITE_CONTACTS",
+  "WRITE_GSERVICES",
+  "WRITE_SECURE_SETTINGS",
+  "WRITE_SETTINGS",
+  "WRITE_VOICEMAIL",
+];
+
