+++ expo-quest/plugin/src/withQuest.ts
@@ -1,5 +1,10 @@
-import { ConfigPlugin, withGradleProperties, withAndroidManifest } from '@expo/config-plugins';
-import { withProhibitedPermissions } from './withProhibitedPermissions';
+import {
+  ConfigPlugin,
+  withGradleProperties,
+  withAndroidManifest,
+} from "@expo/config-plugins";
+import { withProhibitedPermissions } from "./withProhibitedPermissions";
+import withCustomAndroidManifest from "./withCustomAndroidManifest";
 
 type HorizonOptions = {
   questAppId?: string;
@@ -9,15 +14,20 @@ type HorizonOptions = {
   disableVrHeadtracking?: boolean;
 };
 
-const withQuest: ConfigPlugin<HorizonOptions> = (config, options = {}) => {
+const USE_EXPERIMENTAL_PLUGIN = true;
 
+const withQuest: ConfigPlugin<HorizonOptions> = (config, options = {}) => {
   if (process.env.EXPO_HORIZON) {
+    if (USE_EXPERIMENTAL_PLUGIN) {
+      config = withCustomAndroidManifest(config, options);
+    } else {
       config = withQuestEnabled(config);
       config = withQuestAppId(config, options);
       config = withPanelSize(config, options);
       config = withSupportedDevices(config, options);
       config = withVrHeadtracking(config, options);
       config = withProhibitedPermissions(config);
+    }
   }
 
   return config;
@@ -26,9 +36,9 @@ const withQuest: ConfigPlugin<HorizonOptions> = (config, options = {}) => {
 const withQuestEnabled: ConfigPlugin = (config) => {
   return withGradleProperties(config, (config) => {
     config.modResults.push({
-      type: 'property',
-      key: 'horizonEnabled',
-      value: 'true',
+      type: "property",
+      key: "horizonEnabled",
+      value: "true",
     });
 
     return config;
@@ -37,11 +47,11 @@ const withQuestEnabled: ConfigPlugin = (config) => {
 
 const withQuestAppId: ConfigPlugin<HorizonOptions> = (config, options = {}) => {
   return withGradleProperties(config, (config) => {
-    const questAppId = options.questAppId ?? '';
+    const questAppId = options.questAppId ?? "";
 
     config.modResults.push({
-      type: 'property',
-      key: 'questAppId',
+      type: "property",
+      key: "questAppId",
       value: questAppId,
     });
 
@@ -56,9 +66,10 @@ const withPanelSize: ConfigPlugin<HorizonOptions> = (config, options = {}) => {
       return config;
     }
 
-    const mainActivity = config.modResults.manifest?.application?.[0]?.activity?.find(
-      (activity: any) => activity.$?.['android:name'] === '.MainActivity'
-    ) as any;
+    const mainActivity =
+      config.modResults.manifest?.application?.[0]?.activity?.find(
+        (activity: any) => activity.$?.["android:name"] === ".MainActivity"
+      ) as any;
 
     if (mainActivity) {
       if (!mainActivity.layout) {
@@ -67,10 +78,10 @@ const withPanelSize: ConfigPlugin<HorizonOptions> = (config, options = {}) => {
 
       const layoutAttrs: any = {};
       if (options.defaultHeight) {
-        layoutAttrs['android:defaultHeight'] = options.defaultHeight;
+        layoutAttrs["android:defaultHeight"] = options.defaultHeight;
       }
       if (options.defaultWidth) {
-        layoutAttrs['android:defaultWidth'] = options.defaultWidth;
+        layoutAttrs["android:defaultWidth"] = options.defaultWidth;
       }
 
       mainActivity.layout.push({
@@ -82,7 +93,10 @@ const withPanelSize: ConfigPlugin<HorizonOptions> = (config, options = {}) => {
   });
 };
 
-const withSupportedDevices: ConfigPlugin<HorizonOptions> = (config, options = {}) => {
+const withSupportedDevices: ConfigPlugin<HorizonOptions> = (
+  config,
+  options = {}
+) => {
   return withAndroidManifest(config, (config) => {
     // Only add meta-data if supportedDevices is explicitly provided
     if (!options.supportedDevices) {
@@ -92,14 +106,14 @@ const withSupportedDevices: ConfigPlugin<HorizonOptions> = (config, options = {}
     const application = config.modResults.manifest?.application?.[0];
 
     if (application) {
-      if (!application['meta-data']) {
-        application['meta-data'] = [];
+      if (!application["meta-data"]) {
+        application["meta-data"] = [];
       }
 
-      application['meta-data'].push({
+      application["meta-data"].push({
         $: {
-          'android:name': 'com.oculus.supportedDevices',
-          'android:value': options.supportedDevices,
+          "android:name": "com.oculus.supportedDevices",
+          "android:value": options.supportedDevices,
         },
       });
     }
@@ -108,7 +122,10 @@ const withSupportedDevices: ConfigPlugin<HorizonOptions> = (config, options = {}
   });
 };
 
-const withVrHeadtracking: ConfigPlugin<HorizonOptions> = (config, options = {}) => {
+const withVrHeadtracking: ConfigPlugin<HorizonOptions> = (
+  config,
+  options = {}
+) => {
   return withAndroidManifest(config, (config) => {
     // Add VR headtracking by default unless explicitly disabled
     if (options.disableVrHeadtracking === true) {
@@ -118,15 +135,15 @@ const withVrHeadtracking: ConfigPlugin<HorizonOptions> = (config, options = {}) 
     const manifest = config.modResults.manifest;
 
     if (manifest) {
-      if (!manifest['uses-feature']) {
-        manifest['uses-feature'] = [];
+      if (!manifest["uses-feature"]) {
+        manifest["uses-feature"] = [];
       }
 
-      manifest['uses-feature'].push({
+      manifest["uses-feature"].push({
         $: {
-          'android:name': 'android.hardware.vr.headtracking',
-          'android:required': 'true',
-          'android:version': '1',
+          "android:name": "android.hardware.vr.headtracking",
+          "android:required": "true",
+          "android:version": "1",
         },
       } as any);
     }
